Moved data access functions to generic data access functions since they are dimension independent
