+++ core/specfem/assembly/fields/data_access.hpp
@@ -5,8 +5,8 @@
  * @brief Data access functions for field operations
  */
 
-#include "dim2/add_on_device.hpp"
-#include "dim2/atomic_add_on_device.hpp"
-#include "dim2/load_on_device.hpp"
-#include "dim2/store_on_device.hpp"
-#include "dim2/store_on_host.hpp"
+#include "data_access/add_on_device.hpp"
+#include "data_access/atomic_add_on_device.hpp"
+#include "data_access/load_on_device.hpp"
+#include "data_access/store_on_device.hpp"
+#include "data_access/store_on_host.hpp"
